fix: refactor standard reports shell, dynamic column configuration, data fetching hooks, and preview panel UI
